ftsa/requirements.txt
gmpy2
pycryptodome
ecdsa
numpy
matplotlib